requirements.txt
aiogram==3.4.1
aiosqlite==0.19.0
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1
bech32==1.2.0
base58==2.1.1
web3==6.15.1
eth-account==0.10.0
eth-keyfile==0.6.0
cryptography==42.0.5

